moving tests to new test file
